# frontend/lib/api/openapi.json
{
  "components": {
    "schemas": {
...
      "BarRows": {
        "properties": {
          "columns": {
            "items": {
              "type": "string"
            },
            "title": "Columns",
            "type": "array"
          },
          "rows": {
            "items": {
              "items": {
                "anyOf": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "type": "array"
            },
            "title": "Rows",
            "type": "array"
          }
        },
        "required": [
          "columns",
          "rows"
        ],
        "title": "BarRows",
        "type": "object"
      },
...
      "BookPayload": {
        "properties": {
          "asks": {
            "items": {
              "items": {
                "type": "number"
              },
              "maxItems": 2,
              "minItems": 2,
              "type": "array"
            },
            "title": "Asks",
            "type": "array"
          },
          "bids": {
            "items": {
              "items": {
                "type": "number"
              },
              "maxItems": 2,
              "minItems": 2,
              "type": "array"
            },
            "title": "Bids",
            "type": "array"
          },
          "imbalance_l1": {
            "title": "Imbalance L1",
            "type": "number"
          },
          "imbalance_w": {
            "title": "Imbalance W",
            "type": "number"
          },
          "levels": {
            "title": "Levels",
            "type": "integer"
          },
          "liquidity_10bps": {
            "title": "Liquidity 10Bps",
            "type": "number"
          },
          "microprice": {
            "title": "Microprice",
            "type": "number"
          },
          "mid": {
            "title": "Mid",
            "type": "number"
          },
          "ofi_1s": {
            "title": "Ofi 1S",
            "type": "number"
          },
          "profile": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DepthProfile"
              },
              {
                "type": "null"
              }
            ]
          },
          "spread_bps": {
            "title": "Spread Bps",
            "type": "number"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          },
          "ts_ms": {
            "title": "Ts Ms",
            "type": "integer"
          },
          "update_id": {
            "title": "Update Id",
            "type": "integer"
          }
        },
        "required": [
          "symbol",
          "ts_ms",
          "bids",
          "asks",
          "mid",
          "microprice",
          "spread_bps",
          "imbalance_l1",
          "imbalance_w",
          "ofi_1s",
          "liquidity_10bps",
          "levels",
          "update_id"
        ],
        "title": "BookPayload",
        "type": "object"
      },
...
      "FeatureCatalogEntry": {
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "group": {
            "title": "Group",
            "type": "string"
          },
          "kind": {
            "enum": [
              "price",
              "bps",
              "ratio",
              "rate",
              "quantity",
              "zscore",
              "categorical",
              "probability"
            ],
            "title": "Kind",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "ops": {
            "items": {
              "type": "string"
            },
            "title": "Ops",
            "type": "array"
          },
          "unit": {
            "title": "Unit",
            "type": "string"
          },
          "values": {
            "items": {
              "type": "string"
            },
            "title": "Values",
            "type": "array"
          }
        },
        "required": [
          "name",
          "label",
          "unit",
          "kind",
          "group",
          "description",
          "ops",
          "values"
        ],
        "title": "FeatureCatalogEntry",
        "type": "object"
      },
...
      "FeaturesPayload": {
        "properties": {
          "ask_qty": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ask Qty"
          },
          "best_ask": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Best Ask"
          },
          "best_bid": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Best Bid"
          },
          "bid_qty": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bid Qty"
          },
          "book_slope_ask": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Book Slope Ask"
          },
          "book_slope_bid": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Book Slope Bid"
          },
          "buy_pressure": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Buy Pressure"
          },
          "imbalance_l1": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Imbalance L1"
          },
          "imbalance_w": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Imbalance W"
          },
          "imbalance_z": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Imbalance Z"
          },
          "last_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Price"
          },
          "liquidity_10bps": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Liquidity 10Bps"
          },
          "liquidity_25bps": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Liquidity 25Bps"
          },
          "liquidity_5bps": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Liquidity 5Bps"
          },
          "liquidity_z": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Liquidity Z"
          },
          "microprice": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Microprice"
          },
          "microprice_dev_bps": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Microprice Dev Bps"
          },
          "mid": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mid"
          },
          "ofi_1s": {
            "title": "Ofi 1S",
            "type": "number"
          },
          "ofi_30s": {
            "title": "Ofi 30S",
            "type": "number"
          },
          "ofi_5s": {
            "title": "Ofi 5S",
            "type": "number"
          },
          "ofi_z": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ofi Z"
          },
          "p_down": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "P Down"
          },
          "p_up": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "P Up"
          },
          "price_vs_vwap_bps": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Vs Vwap Bps"
          },
          "regime": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Regime"
          },
          "regime_direction": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Regime Direction"
          },
          "regime_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Regime Source"
          },
          "session_change_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Change Pct"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "spread_bps": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spread Bps"
          },
          "spread_z": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spread Z"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          },
          "trade_count_30s": {
            "title": "Trade Count 30S",
            "type": "integer"
          },
          "trend": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trend"
          },
          "ts_ms": {
            "title": "Ts Ms",
            "type": "integer"
          },
          "twap": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Twap"
          },
          "velocity": {
            "title": "Velocity",
            "type": "number"
          },
          "velocity_baseline": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Velocity Baseline"
          },
          "velocity_z": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Velocity Z"
          },
          "vol_z": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vol Z"
          },
          "volatility_bps": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volatility Bps"
          },
          "volume_30s": {
            "title": "Volume 30S",
            "type": "number"
          },
          "vwap": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vwap"
          },
          "warmed_up": {
            "title": "Warmed Up",
            "type": "boolean"
          }
        },
        "required": [
          "symbol",
          "ts_ms",
          "source",
          "warmed_up",
          "last_price",
          "mid",
          "microprice",
          "microprice_dev_bps",
          "vwap",
          "twap",
          "price_vs_vwap_bps",
          "session_change_pct",
          "best_bid",
          "best_ask",
          "bid_qty",
          "ask_qty",
          "spread_bps",
          "imbalance_l1",
          "imbalance_w",
          "liquidity_5bps",
          "liquidity_10bps",
          "liquidity_25bps",
          "book_slope_bid",
          "book_slope_ask",
          "ofi_1s",
          "ofi_5s",
          "ofi_30s",
          "velocity",
          "velocity_baseline",
          "buy_pressure",
          "volume_30s",
          "trade_count_30s",
          "volatility_bps",
          "spread_z",
          "velocity_z",
          "vol_z",
          "ofi_z",
          "imbalance_z",
          "liquidity_z",
          "regime",
          "regime_direction",
          "regime_source",
          "trend"
        ],
        "title": "FeaturesPayload",
        "type": "object"
      },
...
      "MarketStatsResponse": {
        "properties": {
          "alerts": {
            "$ref": "#/components/schemas/AlertEngineStats"
          },
          "backtests_running": {
            "items": {
              "type": "integer"
            },
            "title": "Backtests Running",
            "type": "array"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "symbols": {
            "additionalProperties": {
              "$ref": "#/components/schemas/SymbolStats"
            },
            "title": "Symbols",
            "type": "object"
          },
          "timestamp": {
            "title": "Timestamp",
            "type": "string"
          },
          "writer": {
            "$ref": "#/components/schemas/WriterStats"
          },
          "ws": {
            "$ref": "#/components/schemas/WSStats"
          }
        },
        "required": [
          "source",
          "symbols",
          "writer",
          "alerts",
          "backtests_running",
          "ws",
          "timestamp"
        ],
        "title": "MarketStatsResponse",
        "type": "object"
      },
...
      "Problem": {
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Detail"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FieldError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Errors"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Instance"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Request Id"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status"
        ],
        "title": "Problem",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "description": "Real-time market-microstructure intelligence: streaming features, order-book analytics, calibrated ML signals, alerts and backtesting.",
    "title": "AlgoViz",
    "version": "3.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/market/bars": {
      "get": {
        "operationId": "get_bars_api_v1_market_bars_get",
        "parameters": [
          {
            "description": "Symbol; defaults to the primary symbol",
            "in": "query",
            "name": "symbol",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 20,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Symbol; defaults to the primary symbol",
              "title": "Symbol"
            }
          },
          {
            "in": "query",
            "name": "start_ms",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Ms"
            }
          },
          {
            "in": "query",
            "name": "end_ms",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Ms"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 600,
              "maximum": 7200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BarRows"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation Error"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          }
        },
        "summary": "1-second bars (in-memory ring or history)",
        "tags": [
          "Market Data"
        ]
      }
    },
    "/api/v1/market/book": {
      "get": {
        "operationId": "get_book_api_v1_market_book_get",
        "parameters": [
          {
            "description": "Symbol; defaults to the primary symbol",
            "in": "query",
            "name": "symbol",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 20,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Symbol; defaults to the primary symbol",
              "title": "Symbol"
            }
          },
          {
            "in": "query",
            "name": "depth",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Depth",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/BookPayload"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Get Book Api V1 Market Book Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation Error"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          }
        },
        "summary": "L2 order book (top N)",
        "tags": [
          "Market Data"
        ]
      }
    },
    "/api/v1/market/feature-catalog": {
      "get": {
        "operationId": "get_feature_catalog_api_v1_market_feature_catalog_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FeatureCatalogEntry"
                  },
                  "title": "Response Get Feature Catalog Api V1 Market Feature Catalog Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          }
        },
        "summary": "Feature registry (names, units, allowed operators)",
        "tags": [
          "Market Data"
        ]
      }
    },
    "/api/v1/market/features": {
      "get": {
        "operationId": "get_features_api_v1_market_features_get",
        "parameters": [
          {
            "description": "Symbol; defaults to the primary symbol",
            "in": "query",
            "name": "symbol",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 20,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Symbol; defaults to the primary symbol",
              "title": "Symbol"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeaturesPayload"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation Error"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          }
        },
        "summary": "Current streaming features",
        "tags": [
          "Market Data"
        ]
      }
    },
    "/api/v1/market/regime": {
      "get": {
        "operationId": "get_regime_api_v1_market_regime_get",
        "parameters": [
          {
            "description": "Symbol; defaults to the primary symbol",
            "in": "query",
            "name": "symbol",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 20,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Symbol; defaults to the primary symbol",
              "title": "Symbol"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegimePayload"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation Error"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          }
        },
        "summary": "Current regime state",
        "tags": [
          "Market Data"
        ]
      }
    },
    "/api/v1/market/stats": {
      "get": {
        "operationId": "get_stats_api_v1_market_stats_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketStatsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          }
        },
        "summary": "Engine statistics (all symbols)",
        "tags": [
          "Market Data"
        ]
      }
    },
    "/api/v1/market/symbols": {
      "get": {
        "operationId": "get_symbols_api_v1_market_symbols_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SymbolsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          }
        },
        "summary": "Active symbols and allowlist",
        "tags": [
          "Market Data"
        ]
      }
    },
    "/api/v1/market/trades": {
      "get": {
        "operationId": "get_trades_api_v1_market_trades_get",
        "parameters": [
          {
            "description": "Symbol; defaults to the primary symbol",
            "in": "query",
            "name": "symbol",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 20,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Symbol; defaults to the primary symbol",
              "title": "Symbol"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TradePayload"
                  },
                  "title": "Response Get Trades Api V1 Market Trades Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation Error"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          }
        },
        "summary": "Recent trades",
        "tags": [
          "Market Data"
        ]
      }
    },
...
  }
}